refactor: share checkArgumrnts and settings
